+++ saes/utils.py
@@ -1,0 +1,18 @@
+import torch
+from EWOthello.mingpt.model import GPTConfig, GPTforProbing
+
+device='cuda' if torch.cuda.is_available() else 'cpu'
+
+
+def load_pre_trained_gpt(probe_path, probe_layer):
+    """
+    loads the model at probe_path and wires it to run through probe_layer
+    """
+    n_layer = int(probe_path[-5:-4])
+    n_head = int(probe_path[-3:-2])
+    mconf = GPTConfig(61, 59, n_layer=n_layer, n_head=n_head, n_embd=512)
+    GPT_probe = GPTforProbing(mconf, probe_layer)
+    
+    GPT_probe.load_state_dict(torch.load(probe_path + f"GPT_Synthetic_{n_layer}Layers_{n_head}Heads.ckpt", map_location=device))
+    GPT_probe.eval()
+    return GPT_probe
